+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.1.0
+Run `--help` to see all options
+Google gemini added, run `ls` to see models and `set` to use
+Split out the streaming functions into own files, the single cli.js was getting too long
+
 ## 1.0.1
 Docs added and this changelog